requirements.txt
httpx
lxml
schedule
python-dotenv